LayoutTests/http/tests/xmlhttprequest/sync-delegate-callbacks-expected.txt

ALERT: sending synchronous xhr
ALERT: hello
ALERT: sending synchronous xhr
ALERT: hello
